requirements.txt
aiohttp==3.10.5
Pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.0.1
